+++ theoriq/api/v1alpha1/schemas/__init__.py
@@ -1,3 +1,4 @@
 from .agent import AgentResponse
 from .challenge import ChallengeRequestBody, ChallengeResponseBody
 from .request import ExecuteRequestBody
+from .metrics import MetricsRequestBody
