+++ theoriq/biscuit/utils.py
@@ -20,3 +20,19 @@ def hash_public_key(key: PublicKey) -> str:
 def get_new_key_pair() -> Tuple[str, str]:
     kp = KeyPair()
     return f"0x{kp.public_key.to_hex()}", f"0x{kp.private_key.to_hex()}"
+
+
+def verify_address(address: str) -> str:
+    """
+    Verify the address
+    :raise TypeError: if the address is not 32 bytes long or does not only contain hex digits
+    """
+    add = address.removeprefix("0x").strip()
+    try:
+        length = len(bytes.fromhex(add))
+        if length != 32:
+            raise TypeError(f"address must be 32 bytes long: {address}")
+    except ValueError as e:
+        raise TypeError(f"address must only contain hex digits: {address}") from e
+    else:
+        return add
