+++ theoriq/extra/flask/v1alpha2/flask.py
@@ -12,14 +12,15 @@ from theoriq.agent import Agent, AgentConfig
 from theoriq.api import ExecuteContextV1alpha2, ExecuteRequestFnV1alpha2
 from theoriq.api.v1alpha2.schemas import ExecuteRequestBody as ExecuteRequestBodyV1alpha2
 from theoriq.biscuit import AgentAddress, TheoriqBiscuitError
-from theoriq.extra.flask.common import (
+from theoriq.extra.globals import agent_var
+
+from ..common import (
     add_biscuit_to_response,
     build_error_payload,
     new_error_response,
     process_biscuit_request,
     theoriq_system_blueprint,
 )
-from theoriq.extra.globals import agent_var
 
 logger = logging.getLogger(__name__)
